requirements.txt
python-telegram-bot==20.5
pytesseract==0.3.10
Pillow==12.0.0
gspread==5.8.0
oauth2client==4.1.3
pandas==2.2.3
openpyxl==3.1.2
requests==2.31.0
